Move suite implementation into main flow file, add some typings.
